% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBRKC}
\alias{BBRKC}
\title{Bristol Bay Red King Crab ESP data}
\format{
A data frame with 42 rows and 23 variables:
\describe{
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{PCOD_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{BENPRED_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{BENINVERT_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{RACE_CP}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_CP}}{double COLUMN_DESCRIPTION}
\item{\code{RACE_SUM_BT}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_SUM_BT}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_SPR_BT}}{double COLUMN_DESCRIPTION}
\item{\code{RKC_REC_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{RKC_MMB}}{double COLUMN_DESCRIPTION}
\item{\code{PDO}}{double COLUMN_DESCRIPTION}
\item{\code{WIND_STRESS}}{double COLUMN_DESCRIPTION}
\item{\code{CHL_A}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_OA}}{double COLUMN_DESCRIPTION}
\item{\code{PCOD_OVLP}}{double COLUMN_DESCRIPTION}
\item{\code{MALE_LW_RESID}}{double COLUMN_DESCRIPTION}
\item{\code{SOCK_ABUN}}{double COLUMN_DESCRIPTION}
\item{\code{CATCH_DISTSHORE}}{double COLUMN_DESCRIPTION}
\item{\code{RKC_MIEW}}{double COLUMN_DESCRIPTION}
\item{\code{MATMALE_D95}}{double COLUMN_DESCRIPTION}
\item{\code{MATFEM_D95}}{double COLUMN_DESCRIPTION}
\item{\code{AO}}{double COLUMN_DESCRIPTION}
}
}
\usage{
BBRKC
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
