% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{check_data}
\alias{check_data}
\title{Check ESP data columns}
\usage{
check_data(data, fill = TRUE)
}
\arguments{
\item{data}{The ESP data}

\item{fill}{Whether to add missing columns and fill with NAs}
}
\value{
A tibble
}
\description{
This function checks that the ESP data has the expected columns.
If any columns are missing, they are filled with NAs.
}
