% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indicator_table.R
\name{create_indicator_table}
\alias{create_indicator_table}
\title{Create a status table for the short-form summary report}
\usage{
create_indicator_table(
  data,
  dir,
  type,
  img_w = 4.5,
  img_h = 2.5,
  colnames = c("Indicator", "2024 Status", "Trend")
)
}
\arguments{
\item{data}{The template .csv}

\item{dir}{The directory where images referenced in the .csv are saved}

\item{type}{The type of template. One of c("Ecosystem", "Socioeconomic")}

\item{img_w}{The width of the time series plot (inches)}

\item{img_h}{The height of the time series plot (inches)}

\item{colnames}{The column names for the table. Defaults to c("Indicator", "2024 Status", "Trend").}
}
\description{
This function creates a status table for the short-form summary report.
}
