% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_template}
\alias{create_template}
\title{Create an ESP template}
\usage{
create_template(path = getwd(), type = "full")
}
\arguments{
\item{path}{Where to create the template. Defaults to the present working directory.}

\item{type}{The type of template. One of c("full", "partial", "report card", "one pager")}
}
\description{
This function creates an ESP template at the specified path. DEPRECATED -- PLAN TO REMOVE
}
