% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebs_pcod}
\alias{ebs_pcod}
\title{Eastern Bering Sea Pacific Cod ESP data}
\format{
A data frame with 44 rows and 19 variables:
\describe{
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{North_Pacific_Index_Winter_Spring}}{double COLUMN_DESCRIPTION}
\item{\code{Sea_Ice_Extent_Advance_DJF}}{double COLUMN_DESCRIPTION}
\item{\code{Sea_Ice_Extent_Retreat_MAM}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Summer_Surface_Temperature_EBS}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Bottom_Temperature_EBS_Shelf_ROMS}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Bloom_Timing_SEBS}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Euphausiid_Abundance_EBS}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Condition_Juveniles_EBS_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Condition_Adults_EBS_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Center_Gravity_Eastings_EBS_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Center_Gravity_Northings_EBS_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Area_Occupied_EBS_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Arrowtooth_Flounder_Total_Biomass_Assessment_BSAI}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Real_Exvessel_Value_EBS_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Real_Exvessel_Price_EBS_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Real_Revenue_Per_Unit_Effort_EBS_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Regional_Quotient_Processing_Revenue_Dutch_Harbor_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Regional_Quotient_Harvesting_Revenue_Dutch_Harbor_Fishery}}{double COLUMN_DESCRIPTION}
}
}
\usage{
ebs_pcod
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
