% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{esp_combo_score}
\alias{esp_combo_score}
\title{Plot a figure of combo ESP scores}
\usage{
esp_combo_score(data, species, region, out = "ggplot", name, ...)
}
\arguments{
\item{data}{The ESP indicator data (LONG format).}

\item{species}{The species name}

\item{region}{The stock region}

\item{out}{Whether the function should save the plot, or return a ggplot object. One of c("ggplot", "save")}

\item{name}{The file name for the image. Will be saved relative to the working directory. Only needed if saving the plot.}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
An image file
}
\description{
This function is a alternative combo plot of the overall ESP scores over time.
}
