% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R, R/figure_functions_long.R
\name{esp_metrics}
\alias{esp_metrics}
\title{Plot a metric panel figure}
\usage{
esp_metrics(
  data,
  species,
  region,
  approved = TRUE,
  order = FALSE,
  out,
  name,
  ...
)

esp_metrics(
  data,
  species,
  region,
  approved = TRUE,
  order = FALSE,
  out,
  name,
  ...
)
}
\arguments{
\item{data}{The ESP metric panel data (suggest to use \code{AKesp::metric_panel}).}

\item{species}{The species name}

\item{region}{The stock region}

\item{approved}{Boolean. Whether to use only approved metrics (\code{approved = TRUE})}

\item{order}{Boolean. Whether to order the metrics by value (\code{order = TRUE}). Else will be in standard order.}

\item{out}{Whether the function should save the plot, or return a ggplot object. One of c("ggplot", "save")}

\item{name}{The file name for the image. Will be saved relative to the working directory.}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
An image file

An image file
}
\description{
This function plots an ESP metric panel

This function plots an ESP metric panel
}
