% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{esp_stock_options}
\alias{esp_stock_options}
\title{View AKFIN ESP stock options}
\usage{
esp_stock_options()
}
\value{
A tibble
}
\description{
This function shows which stocks have data available on the AKFIN web service. It does not take any parameters.
}
