% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{esp_traffic}
\alias{esp_traffic}
\title{Plot indicator traffic light figure}
\usage{
esp_traffic(data, name, out, ...)
}
\arguments{
\item{data}{The ESP indicator data. Should have a column for Year and a column for each indicator.}

\item{name}{The file name for the image. Will be saved relative to the working directory.}

\item{out}{Whether the function should save the plot, or return a ggplot object. One of c("ggplot", "save")}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
An image file
}
\description{
This function plots ESP indicator traffic light figures
}
