% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions_long.R
\name{esp_traffic_long}
\alias{esp_traffic_long}
\title{Plot indicator traffic light figure}
\usage{
esp_traffic_long(
  data,
  name,
  out = "ggplot",
  paginate = FALSE,
  label = TRUE,
  caption = "",
  ncolumn = 1,
  ...
)
}
\arguments{
\item{data}{The ESP indicator data. Should have a column for Year and a column for each indicator.}

\item{name}{The file name for the image. Will be saved relative to the working directory.}

\item{out}{Whether the function should save the plot, or return a ggplot object. One of c("ggplot", "save")}

\item{paginate}{Whether to paginate the plots with \code{ggforce::facet_wrap_paginate}}

\item{label}{Whether to label the facets}

\item{ncolumn}{How many columns the figure should have (1 by default)}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
An image file
}
\description{
This function plots ESP indicator traffic light figures
}
