% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{esp_traffic_tab_long}
\alias{esp_traffic_tab_long}
\title{Create indicator traffic light table - LONG data}
\usage{
esp_traffic_tab_long(data, year, cap = "Traffic light scoring")
}
\arguments{
\item{data}{The ESP indicator data. Should have a column for Year and a column for each indicator.}

\item{year}{The year(s) to use for the traffic light analysis. Either a single number or a numeric vector.}

\item{cap}{The table caption.}
}
\value{
A flextable
}
\description{
This function creates an ESP indicator traffic light table
}
