% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_pcod}
\alias{goa_pcod}
\title{Gulf of Alaska Pacific Cod ESP data}
\format{
A data frame with 44 rows and 24 variables:
\describe{
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{Marine_Heatwave_Cumulative_Intensity_Spawning}}{double COLUMN_DESCRIPTION}
\item{\code{Spawning_Habitat_Suitability_GOA}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Bottom_Temperature_GOA_Shelf_CFSR_20-40cm}}{double COLUMN_DESCRIPTION}
\item{\code{Eddy_Kinetic_Energy_Kodiak_Area_D}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Bloom_Timing_WCGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Copepods_YOY_Shelikof}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Euphausiid_Abundance_Kodiak}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_CPUE_Larvae_Shelikof}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_CPUE_YOY_Kodiak}}{double COLUMN_DESCRIPTION}
\item{\code{Common_Murre _Reproductive_Success_Chowiet}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Condition_Juveniles_GOA_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Condition_Adults_GOA_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Center_Gravity_NorthEastings_GOA_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Area_Occupied_GOA_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Steller_Sea_Lions_Adults}}{double COLUMN_DESCRIPTION}
\item{\code{Arrowtooth_Flounder_Total_Biomass_Assessment_GOA}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Real_Exvessel_Value_GOA_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Real_Exvessel_Price_GOA_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Real_Revenue_Per_Unit_Effort_GOA_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Regional_Quotient_Processing_Revenue_Kodiak_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Regional_Quotient_Harvesting_Revenue_Kodiak_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Regional_Quotient_Processing_Revenue_Small_Communities_GOA_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pacific_Cod_Regional_Quotient_Harvesting_Revenue_Small_Communities_GOA_Fishery}}{double COLUMN_DESCRIPTION}
}
}
\usage{
goa_pcod
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
