% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_pollock}
\alias{goa_pollock}
\title{Gulf of Alaska Pollock ESP data}
\format{
A data frame with 44 rows and 34 variables:
\describe{
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Heatwave_GOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Surface_Temperature_WCGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Bottom_Temperature_WCGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Wind_Direction_CGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Chlorophyll_Concentration_WCGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Bloom_Timing_WCGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Copepods_Larvae_Shelikof}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Copepods_YOY_Shelikof}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Euphausiid_Abundance_Kodiak}}{double COLUMN_DESCRIPTION}
\item{\code{Auklet_Reproductive_Success_Chowiet}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Pollock_CPUE_Larvae_Shelikof}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Pollock_CPUE_YOY_Shelikof}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Pollock_Condition_YOY_Shelikof}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Pollock_CPUE_YOY_Kodiak}}{double COLUMN_DESCRIPTION}
\item{\code{Pollock_Relative_Biomass_YOY_Aiktak_Tufted_Puffin}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Pollock_Predation_Age_1}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Pollock_Euphausiid_Diet_Juveniles}}{double COLUMN_DESCRIPTION}
\item{\code{Fall_Pollock_Condition_Adult_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Winter_Pollock_Condition_Adult_Acoustic}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Pollock_Center_Gravity_Northeast_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Pollock_Area_Occupied_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Arrowtooth_Biomass_Assessment}}{double COLUMN_DESCRIPTION}
\item{\code{Pacific_ocean_perch_Biomass_Assessment}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Biomass_Assessment}}{double COLUMN_DESCRIPTION}
\item{\code{Steller_Sea_Lions_Adults}}{double COLUMN_DESCRIPTION}
\item{\code{Winter_Spring_Pollock_CPUE_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Fall_Pollock_CPUE_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pollock_Real_Exvessel_Price_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Winter_Spring_Pollock_Roe_Per_Unit_Catch_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pollock_Regional_Quotient_Processing_Revenue_Kodiak_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pollock_Regional_Quotient_Processing_Revenue_Small_Communities_GOA_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pollock_Regional_Quotient_Harvesting_Revenue_Kodiak_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Pollock_Regional_Quotient_Harvesting_Revenue_Small_Communities_GOA_Fishery}}{double COLUMN_DESCRIPTION}
}
}
\usage{
goa_pollock
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
