% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{join_order}
\alias{join_order}
\title{Join order information to ESP data}
\usage{
join_order(data)
}
\arguments{
\item{data}{The ESP data}
}
\value{
A tibble
}
\description{
This function joins order information to ESP data.
}
