% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_facets}
\alias{label_facets}
\title{Label ggplot facets}
\usage{
label_facets(
  p,
  open = "(",
  close = ")",
  tag_pool = letters,
  x = -Inf,
  y = Inf,
  hjust = -0.5,
  vjust = 1.5,
  fontface = 2,
  family = "",
  ...
)
}
\arguments{
\item{p}{A faceted ggplot.}

\item{open}{Opening character}

\item{close}{Closing character}

\item{tag_pool}{Character vector of labels}

\item{x}{x position within panel}

\item{y}{y position within panel}

\item{hjust}{hjust}

\item{vjust}{vjust}

\item{fontface}{fontface}

\item{family}{font family}

\item{...}{passed to \code{ggplot2::geom_text}}
}
\value{
A ggplot
}
\description{
This function labels gggplot facets with letters. Modified from \code{egg::tag_facet()}
}
