% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_status}
\alias{label_status}
\title{Label indicator status on ggplot}
\usage{
label_status(
  p,
  tag_pool = letters,
  x = -Inf,
  y = Inf,
  hjust = -30,
  vjust = 1.5,
  fontface = 2,
  family = "",
  ...
)
}
\arguments{
\item{p}{A faceted ggplot.}

\item{tag_pool}{Character vector of labels}

\item{x}{x position within panel}

\item{y}{y position within panel}

\item{hjust}{hjust}

\item{vjust}{vjust}

\item{fontface}{fontface}

\item{family}{font family}

\item{...}{passed to \code{ggplot2::geom_text}}
}
\value{
A ggplot
}
\description{
This function labels a ggplot with recent indicator status. Modified from \code{egg::tag_facet()}
}
