% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_indicators}
\alias{list_indicators}
\title{Create an indicator list}
\usage{
list_indicators(data, indicator_type)
}
\arguments{
\item{data}{The indicator data}

\item{indicator_type}{One of \code{c("Ecosystem", "Socioeconomic")}}
}
\description{
This function creates an indicator list from the data.
}
