% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card_functions.R
\name{make_traffic_plot}
\alias{make_traffic_plot}
\title{Render the figure section}
\usage{
make_traffic_plot(dat)
}
\arguments{
\item{dat}{The indicator data}
}
\description{
This function renders an ESP figure.
}
