% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metric_panel}
\alias{metric_panel}
\title{Metric panel data}
\format{
A data frame with 476 rows and 20 variables:
\describe{
\item{\code{Stock}}{character COLUMN_DESCRIPTION}
\item{\code{Group}}{character COLUMN_DESCRIPTION}
\item{\code{Region}}{character COLUMN_DESCRIPTION}
\item{\code{Metric}}{character COLUMN_DESCRIPTION}
\item{\code{Category}}{character COLUMN_DESCRIPTION}
\item{\code{Use}}{character COLUMN_DESCRIPTION}
\item{\code{Threshold.Value.High}}{double COLUMN_DESCRIPTION}
\item{\code{Threshold.Value.Low}}{double COLUMN_DESCRIPTION}
\item{\code{Reverse}}{character COLUMN_DESCRIPTION}
\item{\code{Mean}}{double COLUMN_DESCRIPTION}
\item{\code{Stdev}}{double COLUMN_DESCRIPTION}
\item{\code{Max}}{double COLUMN_DESCRIPTION}
\item{\code{Min}}{double COLUMN_DESCRIPTION}
\item{\code{Value}}{double COLUMN_DESCRIPTION}
\item{\code{Quality}}{double COLUMN_DESCRIPTION}
\item{\code{Order1}}{integer COLUMN_DESCRIPTION}
\item{\code{Order2}}{integer COLUMN_DESCRIPTION}
\item{\code{Rank0}}{character COLUMN_DESCRIPTION}
\item{\code{Flag}}{integer COLUMN_DESCRIPTION}
\item{\code{Rank}}{character COLUMN_DESCRIPTION}
}
}
\usage{
metric_panel
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
