% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{prep_ind_data}
\alias{prep_ind_data}
\title{Prep indicator data}
\usage{
prep_ind_data(data, recent = TRUE, label_width = 50)
}
\arguments{
\item{data}{The data (pulled from AKFIN)}

\item{recent}{Boolean. Whether to calculate the score only for the most recent year.}

\item{label_width}{Width of facet labels (number of characters). Defaults to 50.}
}
\value{
A tibble
}
\description{
This function prepares indicator data. Adds SCORE column.
}
