% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_AKesp_book.R
\name{render_AKesp_book}
\alias{render_AKesp_book}
\title{Shortcut to render the bookdown SOP}
\usage{
render_AKesp_book()
}
\description{
This function renders the bookdown SOP from the files in the \code{book} folder.
}
