% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{render_esp}
\alias{render_esp}
\title{Create an ESP}
\usage{
render_esp(
  out_name = "EXAMPLE-FULL-ESP.docx",
  dir = getwd(),
  num = 1,
  authors = "Kalei Shotwell, Abby Tyrell",
  year = 2021,
  contributors = "Kalei Shotwell, Abby Tyrell",
  fish = "Myfish",
  region = "Myarea",
  fig_spreadsheet = "figure_spreadsheet.csv",
  tab_spreadsheet = "table_spreadsheet.csv",
  ref_spreadsheet = "references_spreadsheet.csv",
  esp_text = "full-esp-text-template.docx",
  esp_type = "full",
  esp_data = NULL,
  con_model_path = "images/con_model.png",
  stock_image = "default"
)
}
\arguments{
\item{out_name}{The file name for the report}

\item{dir}{The directory that the template is in. Defaults to the working directory.}

\item{num}{The appendix number of the EPS}

\item{authors}{The names of the authors, as a single character string}

\item{year}{The year of the ESP}

\item{contributors}{The names of the contributors, as a single character string}

\item{fish}{The name of the fish species}

\item{region}{The name of the stock area}

\item{fig_spreadsheet}{The file path to the filled out figure spreadsheet (one of the template documents)}

\item{tab_spreadsheet}{The file path to the filled out table spreadsheet (one of the template documents)}

\item{ref_spreadsheet}{The file path to the filled out references spreadsheet (one of the template documents)}

\item{esp_text}{The file path to the filled out text template (one of the template documents)}

\item{esp_type}{The type of ESP to make. One of c("full", "partial", "report card", "one pager")}

\item{esp_data}{The data to use for automated analyses. Currently only required for report card ESPs.}

\item{con_model_path}{The path to the conceptual model. Currently only required for report card ESPs.}

\item{stock_image}{The path to an image to use on the cover page. Shows the NOAA logo as a default.}
}
\description{
This function creates an ESP from a template. If left empty, an example report will be created.
}
