% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{render_fig}
\alias{render_fig}
\title{Render the figure section}
\usage{
render_fig(img, lab = "no-name", cap = "no-caption", alt = "no-alt")
}
\arguments{
\item{img}{The file path to a figure}

\item{lab}{The figure chunk label (keyword)}

\item{cap}{The figure caption}

\item{alt}{The figure alt text}
}
\description{
This function renders an ESP figure.
}
