% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{render_ref}
\alias{render_ref}
\title{Create \code{references.bib} from template spreadsheet}
\usage{
render_ref(refs = "references_spreadsheet.csv", dir)
}
\arguments{
\item{refs}{The file path to a reference spreadsheet. Defaults to \code{references_spreadsheet.csv}.}

\item{dir}{The directory where references.bib should be saved. Only needed because of packaging issues.}
}
\description{
This function creates \code{references.bib} from a template spreadsheet
}
