% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card_functions.R
\name{render_rpt_card}
\alias{render_rpt_card}
\title{Create an ESP report card}
\usage{
render_rpt_card(...)
}
\arguments{
\item{...}{Passed to \code{rmarkdown::render}. Set rmarkdown parameter values with \verb{..., params = list(...), ...}. Available parameters and their defaults are specified below.
Available parameters and their defaults:
num: 1
authors: "Kalei Shotwell, Abby Tyrell"
year: 2021
contributors: "Kalei Shotwell, Abby Tyrell"
fish: "Myfish"
region: "Myarea"
stock_image: system.file('images/noaa.jpg', package = 'AKesp')
esp_data: AKesp::get_esp_data("Alaska Sablefish")
con_model_path: system.file("images/noaa.jpg", package = "AKesp")
bayes_path: system.file("images/noaa.jpg", package = "AKesp")
ncolumn: 1
min_year: NULL}
}
\description{
This function creates an ESP report card from a template. If left empty, an example report will be created.
}
