% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{render_tab}
\alias{render_tab}
\title{Render the table section}
\usage{
render_tab(tab, lab = "no-name", cap = "no-caption")
}
\arguments{
\item{tab}{The file path to a table}

\item{lab}{The table chunk label (keyword)}

\item{cap}{The table caption}
}
\description{
This function renders an ESP table.
}
