% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{rpt_card_timeseries}
\alias{rpt_card_timeseries}
\title{Plot indicator time series for the report card}
\usage{
rpt_card_timeseries(data, ylab, xlims, new_breaks)
}
\arguments{
\item{data}{The ESP indicator data (LONG format).}

\item{ylab}{The y-axis label}

\item{xlims}{The x-axis limits}

\item{new_breaks}{The x-axis breaks}
}
\value{
A ggplot
}
\description{
This function plots an indicator time series for the report card
}
