% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sablefish}
\alias{sablefish}
\title{Sablefish ESP data}
\format{
A data frame with 44 rows and 29 variables:
\describe{
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Heatwave_GOA}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_250m_Temperature_GOA_Slope}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Surface_Temperature_EGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Surface_Temperature_SEBS}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Chlorophyll_Concentration_EGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Chlorophyll_Concentration_SEBS}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Bloom_Timing_EGOA}}{double COLUMN_DESCRIPTION}
\item{\code{Spring_Peak_Bloom_Timing_SEBS}}{double COLUMN_DESCRIPTION}
\item{\code{Copepod Community Size EGOA Oceanic}}{double COLUMN_DESCRIPTION}
\item{\code{Copepod Community Size WGOA Oceanic}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Euphausiid_Abundance_Kodiak}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Growth_YOY_Middleton_Auklets}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Sablefish_CPUE_Juvenile_WGOA_EAI_ADFG}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Sablefish_CPUE_Juvenile_GOA_BTS_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Spawner_Mean_Age}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Spawner_Age_Evenness}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Sablefish_Condition_Age_4_GOA_LL_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Arrowtooth_Biomass_Assessment}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Catch_Arrowtooth_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Summer_Sablefish_Condition_Adult_GOA_LL_Survey}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_CPUE_Longline_Fishery_GOA}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_CPUE_Pot_Fishery_EBS}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Incidental_Catch_BSAI_Fisheries}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Incidental_Catch_GOA_Fisheries}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Condition_Adult_GOA_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Sablefish_Condition_Adult_BSAI_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Sablefish_Real_Exvessel_Value_Fishery}}{double COLUMN_DESCRIPTION}
\item{\code{Annual_Sablefish_Real_Exvessel_Price_Fishery}}{double COLUMN_DESCRIPTION}
}
}
\usage{
sablefish
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
