% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMBKC}
\alias{SMBKC}
\title{Saint Matthew Island Blue King Crab data}
\format{
A data frame with 41 rows and 24 variables:
\describe{
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{BENINVERT_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{BENPRED_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{PCOD_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{SMBKC_REC_BIO}}{double COLUMN_DESCRIPTION}
\item{\code{SMBKC_MMB}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_OA}}{double COLUMN_DESCRIPTION}
\item{\code{RACE_CP}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_CP}}{double COLUMN_DESCRIPTION}
\item{\code{RACE_SUM_BT}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_SUM_BT}}{double COLUMN_DESCRIPTION}
\item{\code{ROMS_SPR_BT}}{double COLUMN_DESCRIPTION}
\item{\code{PDO}}{double COLUMN_DESCRIPTION}
\item{\code{CHL_A}}{double COLUMN_DESCRIPTION}
\item{\code{WIND_STRESS}}{double COLUMN_DESCRIPTION}
\item{\code{Vessels}}{double COLUMN_DESCRIPTION}
\item{\code{Bycatch}}{double COLUMN_DESCRIPTION}
\item{\code{Total_Potlifts}}{double COLUMN_DESCRIPTION}
\item{\code{CPUE}}{double COLUMN_DESCRIPTION}
\item{\code{Price_per_pound}}{double COLUMN_DESCRIPTION}
\item{\code{Revenue_share}}{double COLUMN_DESCRIPTION}
\item{\code{Processors}}{double COLUMN_DESCRIPTION}
\item{\code{Local_Quotient}}{double COLUMN_DESCRIPTION}
\item{\code{TAC_Utilization}}{double COLUMN_DESCRIPTION}
}
}
\usage{
SMBKC
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
