% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_google_files.R
\name{use_google_files}
\alias{use_google_files}
\title{Use files from Google Drive}
\usage{
use_google_files(url, ...)
}
\arguments{
\item{url}{The URL of the Google folder with the files}

\item{...}{passed to \code{googledrive::drive_download} (ex, \code{overwrite = TRUE})}
}
\description{
This function downloads files from Google Drive to a temporary directory to use when rendering an ESP. Intended for internal use by \code{render_esp}.
}
