% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indicator_table.R
\name{create_indicator_table}
\alias{create_indicator_table}
\title{Create a status table for the short-form summary report}
\usage{
create_indicator_table(data, dir, type)
}
\arguments{
\item{data}{The template .csv}

\item{dir}{The directory where images referenced in the .csv are saved}

\item{type}{The type of template. One of c("Ecosystem", "Socioeconomic")}
}
\description{
This function creates a status table for the short-form summary report.
}
