% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{esp_cor_matrix}
\alias{esp_cor_matrix}
\title{Plot correlation matrices}
\usage{
esp_cor_matrix(data, name, out, ...)
}
\arguments{
\item{data}{The ESP indicator data. LONG format.}

\item{name}{The file name for the image. Will be saved relative to the working directory. Only needed if saving the plot.}

\item{out}{Whether the function should save the plot, or return a ggplot object. One of c("ggplot", "save")}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
An image file or a ggplot
}
\description{
This function creates correlation matrices for ESP indicators.
}
