% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{esp_hist}
\alias{esp_hist}
\title{Plot indicator histograms}
\usage{
esp_hist(data, name, out, ...)
}
\arguments{
\item{data}{The ESP indicator data. Should have a column for Year and a column for each indicator.}

\item{name}{The file name for the image. Will be saved relative to the working directory.}

\item{out}{Whether the function should save the plot, or return a ggplot object. One of c("ggplot", "save")}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
An image file or a ggplot
}
\description{
This function plots ESP indicator hisograms
}
