% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{esp_traffic}
\alias{esp_traffic}
\title{Plot indicator traffic light figure}
\usage{
esp_traffic(
  data,
  name,
  out = "ggplot",
  paginate = FALSE,
  label = TRUE,
  status = FALSE,
  caption = "",
  ncolumn = 1,
  silent = FALSE,
  min_year = NULL,
  chunk_label = "traffic",
  f_units = FALSE,
  y_units = TRUE,
  skip_lines = FALSE,
  ...
)
}
\arguments{
\item{data}{The ESP indicator data. Should have a column for Year and a column for each indicator.}

\item{name}{The file name for the image. Will be saved relative to the working directory.}

\item{out}{Whether the function should save the plot, print a ggplot object in markdown, or return a ggplot object to the working environment (for use with \code{one_pager()}). One of c("ggplot", "save", "one_pager")}

\item{paginate}{Whether to paginate the plots with \code{ggforce::facet_wrap_paginate}}

\item{label}{Whether to label the facets with a, b, c, etc.}

\item{status}{Whether to label the facets with the indicator status}

\item{caption}{A caption for the figure}

\item{ncolumn}{How many columns the figure should have (1 by default)}

\item{silent}{Whether to print the caption}

\item{min_year}{The minimum year to show on the plots. If left NULL (the default), the minimum year will be the first year of the dataset.}

\item{chunk_label}{The label name to look for to create the figure number. This is a work-around to deal with figure pagination.}

\item{f_units}{Whether to add units in the facet header}

\item{y_units}{Whether to add units to the y-axis}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
An image file
}
\description{
This function plots ESP indicator traffic light figures
}
