% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-pager.R
\name{one_pager}
\alias{one_pager}
\title{Create ESP one-pager}
\usage{
one_pager(
  data,
  overall_data,
  years = 2017:2021,
  head_text = "some text",
  foot_text = "more text",
  bayes_path = here::here("bayes.png"),
  header_path = here::here("header.png"),
  output_name = "one-pager.pdf"
)
}
\arguments{
\item{data}{The data you want displayed in traffic light figures/table (pulled from AKFIN)}

\item{overall_data}{The data to use in creating the overall score figure}

\item{years}{The years to display in the traffic light table}

\item{head_text}{The text to display at the top of the one-pager}

\item{foot_text}{The text to display at the bottom of the one-pager}

\item{bayes_path}{The file path to the Bayesian Adaptive Sampling figure (must be a png)}

\item{header_path}{The file path to the image to use as the header (must be a png)}

\item{output_name}{The file name of the output. Include the file extension (can be any file extension that is compatible with \code{ggsave})}
}
\value{
A saved one-pager file
}
\description{
This function creates a one-pager summary of ESP data.
}
