% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{rpt_card_timeseries}
\alias{rpt_card_timeseries}
\title{Plot a time series figure for inclusion in the ESP report card indicator table}
\usage{
rpt_card_timeseries(data, ylab, xlims, new_breaks, type = "Ecosystem")
}
\arguments{
\item{data}{The ESP indicator data -- data for a single indicator time series.}

\item{ylab}{The y-axis label}

\item{xlims}{The x-axis limits}

\item{new_breaks}{The x-axis breaks}

\item{type}{The type of indicator, either "Ecosystem" or "Socioeconomic". Defaults to "Ecosystem".}
}
\value{
A ggplot
}
\description{
This function plots  a time series figure for inclusion in the ESP report card indicator table.
}
